+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.1.12 (unreleased)
+- Add support for 8va treble clefs ('treble_8va') and bridge clefs ('bridge'). Bridge clef acts like percussion clefs - it uses the center staff position as middle C regardless of staff line count.
+- Add support for image transparency with new `opacity` field. The field is accessible through the `Image` constructor and a new `opacity` property.
+
 # 0.1.11 (2023-01-05)
 
 - Fix bug where viewport scales were set and got incorrectly (#89)